requirements.txt
alembic==1.16.5
Flask==3.1.2
Flask_Cors==5.0.0
Flask-Limiter==3.5.0
flask_mail==0.10.0
Flask_Migrate==4.1.0
flask_restx==1.3.2
flask_sqlalchemy==3.1.1
itsdangerous==2.2.0
python-dotenv==1.0.1
Requests==2.32.5
SQLAlchemy==2.0.36
Werkzeug==3.1.4
pytest==8.3.4
pytest-cov==6.0.0
